DEV-556: remove indexclient dependency (#62)
DEV-556: remove indexclient dependency

Removed sharing modules with the client and re-implemented where needed
