+++ .secrets.baseline
@@ -92,14 +92,14 @@
         "hashed_secret": "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 31,
+        "line_number": 23,
         "type": "Basic Auth Credentials"
       },
       {
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 85,
+        "line_number": 222,
         "type": "Secret Keyword"
       }
     ],
