+++ requirements.txt
@@ -4,37 +4,76 @@
 #
 #    pip-compile setup.py
 #
-attrs==19.3.0             # via jsonschema
-cdislogging==1.0.0        # via indexd (setup.py)
-certifi==2020.6.20        # via requests
-chardet==3.0.4            # via requests
-click==7.1.2              # via flask
-configparser==4.0.2       # via importlib-metadata
-contextlib2==0.6.0.post1  # via importlib-metadata, zipp
-git+https://github.com/uc-cdis/doiclient.git@1.0.0#egg=doiclient  # via indexd (setup.py)
-git+https://github.com/uc-cdis/dosclient.git@1.0.0#egg=dosclient  # via indexd (setup.py)
-flask==1.1.2              # via indexd (setup.py)
-# manual edit to install only with py2
-functools32==3.2.3.post2;python_version=="2.7"  # via jsonschema
-future==0.18.2            # via indexd (setup.py)
-idna==2.10                # via requests
-importlib-metadata==1.7.0  # via jsonschema
-git+https://github.com/NCI-GDC/indexclient.git@2.0.0#egg=indexclient  # via indexd (setup.py)
-itsdangerous==1.1.0       # via flask
-jinja2==2.11.2            # via flask
-jsonschema==3.2.0         # via indexd (setup.py)
-markupsafe==1.1.1         # via jinja2
-pathlib2==2.3.5           # via importlib-metadata
-psycopg2==2.8.5           # via indexd (setup.py)
-pyrsistent==0.16.0        # via jsonschema
-requests==2.24.0          # via doiclient, dosclient, indexclient
-scandir==1.10.0           # via pathlib2
-six==1.15.0               # via jsonschema, pathlib2, pyrsistent, sqlalchemy-utils
-sqlalchemy-utils==0.36.3  # via indexd (setup.py)
-sqlalchemy==1.3.18        # via indexd (setup.py), sqlalchemy-utils
-urllib3==1.25.9           # via requests
-werkzeug==0.16.1          # via flask, indexd (setup.py)
-zipp==1.2.0               # via importlib-metadata
+attrs==19.3.0
+    # via jsonschema
+cdislogging==1.0.0
+    # via indexd (setup.py)
+certifi==2020.6.20
+    # via requests
+chardet==3.0.4
+    # via requests
+click==7.1.2
+    # via flask
+configparser==4.0.2
+    # via importlib-metadata
+contextlib2==0.6.0.post1
+    # via
+    #   importlib-metadata
+    #   zipp
+git+https://github.com/uc-cdis/doiclient.git@1.0.0#egg=doiclient
+    # via indexd (setup.py)
+git+https://github.com/uc-cdis/dosclient.git@1.0.0#egg=dosclient
+    # via indexd (setup.py)
+flask==1.1.2
+    # via indexd (setup.py)
+functools32==3.2.3.post2;python_version=="2.7"
+    # via jsonschema
+future==0.18.2
+    # via indexd (setup.py)
+idna==2.10
+    # via requests
+importlib-metadata==1.7.0
+    # via jsonschema
+itsdangerous==1.1.0
+    # via flask
+jinja2==2.11.2
+    # via flask
+jsonschema==3.2.0
+    # via indexd (setup.py)
+markupsafe==1.1.1
+    # via jinja2
+pathlib2==2.3.5
+    # via importlib-metadata
+psycopg2==2.8.5
+    # via indexd (setup.py)
+pyrsistent==0.16.0
+    # via jsonschema
+requests==2.24.0
+    # via
+    #   doiclient
+    #   dosclient
+scandir==1.10.0
+    # via pathlib2
+six==1.15.0
+    # via
+    #   jsonschema
+    #   pathlib2
+    #   pyrsistent
+    #   sqlalchemy-utils
+sqlalchemy-utils==0.36.3
+    # via indexd (setup.py)
+sqlalchemy==1.3.18
+    # via
+    #   indexd (setup.py)
+    #   sqlalchemy-utils
+urllib3==1.25.9
+    # via requests
+werkzeug==0.16.1
+    # via
+    #   flask
+    #   indexd (setup.py)
+zipp==1.2.0
+    # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
